Move structs back into the .c files, making them private.
Added accessors where needed.
