+++ .cirrus.yml
@@ -7,11 +7,12 @@ cirrus-ci_task:
   configure_script:
     - /src/workspace/tools/inject-repo c-toxcore
   test_all_script:
-    - bazel test -k
+    - TEST="bazel test -k
       --build_tag_filters=-haskell
       --test_tag_filters=-haskell
       --remote_download_minimal
       --config=ci
       --config=release
       --config=testnet
-      //c-toxcore/...
+      //c-toxcore/..."
+    - $TEST || $TEST || $TEST || $TEST
