+++ toxcore/DHT.c
@@ -46,6 +46,31 @@
 /* Number of get node requests to send to quickly find close nodes. */
 #define MAX_BOOTSTRAP_TIMES 5
 
+typedef struct DHT_Friend_Callback {
+    dht_ip_cb *ip_callback;
+    void *data;
+    int32_t number;
+} DHT_Friend_Callback;
+
+struct DHT_Friend {
+    uint8_t     public_key[CRYPTO_PUBLIC_KEY_SIZE];
+    Client_data client_list[MAX_FRIEND_CLIENTS];
+
+    /* Time at which the last get_nodes request was sent. */
+    uint64_t    lastgetnode;
+    /* number of times get_node packets were sent. */
+    uint32_t    bootstrap_times;
+
+    /* Symmetric NAT hole punching stuff. */
+    NAT         nat;
+
+    uint16_t lock_count;
+    DHT_Friend_Callback callbacks[DHT_FRIEND_MAX_LOCKS];
+
+    Node_format to_bootstrap[MAX_SENT_NODES];
+    unsigned int num_to_bootstrap;
+};
+
 typedef struct Cryptopacket_Handler {
     cryptopacket_handler_cb *function;
     void *object;
