+++ toxcore/onion_client.h
@@ -182,119 +182,14 @@ void kill_onion_client(Onion_Client *onion_c);
  */
 unsigned int onion_connection_status(const Onion_Client *onion_c);
 
-typedef struct Onion_Node {
-    uint8_t     public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    IP_Port     ip_port;
-    uint8_t     ping_id[ONION_PING_ID_SIZE];
-    uint8_t     data_public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    uint8_t     is_stored;
+typedef struct Onion_Friend Onion_Friend;
 
-    uint64_t    added_time;
+uint16_t onion_get_friend_count(Onion_Client *onion_c);
+Onion_Friend *onion_get_friend(Onion_Client *onion_c, uint16_t friend_num);
+const uint8_t *onion_friend_get_gc_public_key(Onion_Friend *onion_friend);
+void onion_friend_set_gc_public_key(Onion_Friend *onion_friend, const uint8_t *public_key);
 
-    uint64_t    timestamp;
-
-    uint64_t    last_pinged;
-
-    uint8_t     unsuccessful_pings;
-
-    uint32_t    path_used;
-} Onion_Node;
-
-typedef struct Onion_Client_Paths {
-    Onion_Path paths[NUMBER_ONION_PATHS];
-    uint64_t last_path_success[NUMBER_ONION_PATHS];
-    uint64_t last_path_used[NUMBER_ONION_PATHS];
-    uint64_t path_creation_time[NUMBER_ONION_PATHS];
-    /* number of times used without success. */
-    unsigned int last_path_used_times[NUMBER_ONION_PATHS];
-} Onion_Client_Paths;
-
-typedef struct Last_Pinged {
-    uint8_t     public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    uint64_t    timestamp;
-} Last_Pinged;
-
-typedef struct Onion_Friend {
-    uint8_t status; /* 0 if friend is not valid, 1 if friend is valid.*/
-    uint8_t is_online; /* Set by the onion_set_friend_status function. */
-
-    uint8_t know_dht_public_key; /* 0 if we don't know the dht public key of the other, 1 if we do. */
-    uint8_t dht_public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    uint8_t real_public_key[CRYPTO_PUBLIC_KEY_SIZE];
-
-    Onion_Node clients_list[MAX_ONION_CLIENTS];
-    uint8_t temp_public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    uint8_t temp_secret_key[CRYPTO_SECRET_KEY_SIZE];
-
-    uint64_t last_dht_pk_onion_sent;
-    uint64_t last_dht_pk_dht_sent;
-
-    uint64_t last_noreplay;
-
-    uint64_t last_seen;
-
-    Last_Pinged last_pinged[MAX_STORED_PINGED_NODES];
-    uint8_t last_pinged_index;
-
-    recv_tcp_relay_cb *tcp_relay_node_callback;
-    void *tcp_relay_node_callback_object;
-    uint32_t tcp_relay_node_callback_number;
-
-    onion_dht_pk_cb *dht_pk_callback;
-    void *dht_pk_callback_object;
-    uint32_t dht_pk_callback_number;
-
-    uint32_t run_count;
-
-    uint8_t gc_data[GCA_MAX_DATA_LENGTH];
-    uint8_t gc_public_key[ENC_PUBLIC_KEY];
-    int16_t gc_data_length;
-} Onion_Friend;
-
-typedef struct Onion_Data_Handler {
-    oniondata_handler_cb *function;
-    void *object;
-} Onion_Data_Handler;
-
-struct Onion_Client {
-    Mono_Time *mono_time;
-
-    DHT     *dht;
-    Net_Crypto *c;
-    GC_Session *gc_session;
-    Networking_Core *net;
-    Onion_Friend    *friends_list;
-    uint16_t       num_friends;
-
-    Onion_Node clients_announce_list[MAX_ONION_CLIENTS_ANNOUNCE];
-    uint64_t last_announce;
-
-    Onion_Client_Paths onion_paths_self;
-    Onion_Client_Paths onion_paths_friends;
-
-    uint8_t secret_symmetric_key[CRYPTO_SYMMETRIC_KEY_SIZE];
-    uint64_t last_run;
-    uint64_t first_run;
-
-    uint8_t temp_public_key[CRYPTO_PUBLIC_KEY_SIZE];
-    uint8_t temp_secret_key[CRYPTO_SECRET_KEY_SIZE];
-
-    Last_Pinged last_pinged[MAX_STORED_PINGED_NODES];
-
-    Node_format path_nodes[MAX_PATH_NODES];
-    uint16_t path_nodes_index;
-
-    Node_format path_nodes_bs[MAX_PATH_NODES];
-    uint16_t path_nodes_index_bs;
-
-    Ping_Array *announce_ping_array;
-    uint8_t last_pinged_index;
-    Onion_Data_Handler onion_data_handlers[256];
-
-    uint64_t last_packet_recv;
-
-    unsigned int onion_connected;
-    bool udp_connected;
-};
+void onion_friend_set_gc_data(Onion_Friend *onion_friend, const uint8_t *gc_data, int16_t gc_data_length);
+int16_t onion_friend_gc_data_length(Onion_Friend *onion_friend);
 
 #endif
